Fix warning in qmetaobjectbuilder.cpp
src\corelib\kernel\qmetaobjectbuilder.cpp(1489): warning C4267: '+=':
conversion from 'size_t' to 'int', possible loss of data

+++ src/corelib/kernel/qmetaobjectbuilder.cpp
@@ -1486,7 +1486,7 @@ static int buildMetaObject(QMetaObjectBuilderPrivate *d, char *buf,
                 types++;
             }
         }
-        size += sizeof(QMetaType) * d->properties.size();
+        size += static_cast<int>(sizeof(QMetaType) * d->properties.size());
     }
 
     // Align the final size and return it.
